Handle many "uppercase" keys and use a table-driven approach
The "uppercase" symbols are entered by simulating a shift-press,
key-press-release, and shift-release; as if the events came from
an actual keyboard. So far uppercase letters and the symbols
produced by the numeric row with an US keyboard layout are added.
It is trivial to list more allowed combinations in the translation
table.